HACK: Add permissions for neversleep
Because of a prevalent bug concerning suspending to deep sleep on the
tone platform, introduce a workaround.

Hold an infinite wakelock so the device never enters deep sleep and thus
has no problems waking up in a timely manner.

neversleep.rc gets assigned vendor_init and needs to write to
/sys/power/wake_{un,}lock and as such needs capability2 block_suspend.

Affected properties:
- persist.vendor.neversleep=(true|false)
- vendor.neversleep.default=true
- vendor.neversleep.trigger=(true|false)
They get labeled as vendor_neversleep_prop and as such
ExtendedSettings(=system_app), vendor_init and neversleep.sh need to be
able to read and write this prop.

Change-Id: I7fac1093c66bb81b872d22d80d2f96ef7760894f
Signed-off-by: Felix <google@ix5.org>